Added modifications gmenguez did. See http://relavak.wordpress.com/2009/10/17/temper-temperature-sensor-linux-driver/ and http://box.jisko.net/d/2f907e78

+++ temper.h
@@ -33,9 +33,9 @@ typedef struct Temper Temper;
 Temper *TemperCreateFromDeviceNumber(int deviceNum, int timeout, int debug);
 void TemperFree(Temper *t);
 
-int TemperGetTemperatureInC(Temper *t, float *tempC);
+int TemperGetTemperatureInC(Temper *t, double *tempC);
 int TempterGetOtherStuff(Temper *t, char *buf, int length);
 
-int TemperGetTempAndRelHum(Temper *t, float *tempC, float *relhum);
+int TemperGetTempAndRelHum(Temper *t, double *tempC, double *relhum);
 
 #endif
